Refactor: Centralize ElectronAPI interface
I've moved the ElectronAPI interface to src/types.ts to provide a single source of truth for the Electron API contract. This resolves a TypeScript error where 'showSaveDialog' was not found on the type.
